Test compatibility with FIPS-enabled OpenSSL 1.0.2
Some of the standard library crypto/cipher AES-GCM tests fail when using
OpenSSL 1.0.2 in FIPS mode, but pass with the same library when FIPS
mode is disabled. Extend the AES-GCM tests to cover all the failure
cases caught by the standard library tests and build the OpenSSL 1.0.2
library used in CI to be FIPS enabled.

Link the OpenSSL 1.0.2 CI build against FIPS Object Module 2.0.1 rather
than the latest version (2.0.16) as at least one commercially-supported
FIPS validated build of OpenSSL 1.0.2 is known to use that version of
the FIPS Object Module and some of the failures seen with 2.0.1 do not
reproduce with 2.0.16.

Signed-off-by: Cory Snider <csnider@mirantis.com>